@@ -1,22 +1,22 @@
 {
-  "name": "sample-groceries-apollo",
+  "name": "racket-o-meter-frontend",
   "version": "1.0.0",
   "description": "A NativeScript-built iOS and Android app for managing grocery lists with Apollo Stack",
   "repository": {
     "type": "git",
-    "url": "https://github.com/racketometer/sample-groceries-apollo.git"
+    "url": "https://github.com/racketometer/frontend-application.git"
   },
   "keywords": [
     "NativeScript"
   ],
-  "author": "TJ VanToll <tj.vantoll@gmail.com> (https://www.tjvantoll.com/)",
+  "author": "Bjørn Sørensen <https://github.com/crevil>, Jesper O. Christensen <https://github.com/karnich>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/racketometer/sample-groceries-apollo/issues"
+    "url": "https://github.com/racketometer/frontend-application/issues"
   },
-  "homepage": "https://github.com/racketometer/sample-groceries-apollo",
+  "homepage": "https://github.com/racketometer/frontend-application",
   "nativescript": {
-    "id": "org.racketometer.groceries",
+    "id": "com.caveofcode.rom.frontend",
     "tns-android": {
       "version": "2.3.0"
     },
@@ -26,22 +26,25 @@
   },
   "scripts": {
     "tslint": "tslint \"app/**/*.ts\"",
-    "android": "tns livesync android --watch --bundle"
+    "android": "tns livesync android --watch --bundle",
+    "ios": "tns livesync ios --emulator --bundle",
+    "clean": "rimraf platforms/ && rimraf node_modules/ && rimraf lib/ && rimraf hooks"
   },
   "dependencies": {
-    "@angular/common": "2.0.1",
-    "@angular/compiler": "2.0.1",
-    "@angular/core": "2.0.1",
-    "@angular/platform-browser": "2.0.1",
-    "@angular/platform-browser-dynamic": "2.0.1",
-    "@angular/platform-server": "2.0.1",
-    "@angular/router": "3.0.1",
-    "angular2-apollo": "0.4.6",
-    "email-validator": "1.0.4",
-    "graphql": "0.7.1",
+    "@angular/common": "2.1.0",
+    "@angular/compiler": "2.1.0",
+    "@angular/core": "2.1.0",
+    "@angular/platform-browser": "2.1.0",
+    "@angular/platform-browser-dynamic": "2.1.0",
+    "@angular/platform-server": "2.1.0",
+    "@angular/router": "3.1.0",
+    "angular2-apollo": "0.5.0",
+    "apollo-client": "0.4.21",
+    "email-validator": "1.0.7",
+    "graphql": "0.7.2",
     "graphql-tag": "0.1.14",
-    "nativescript-angular": "1.0.1",
-    "nativescript-bluetooth": "1.1.2",
+    "nativescript-angular": "1.1.0",
+    "nativescript-bluetooth": "1.1.3",
     "nativescript-iqkeyboardmanager": "1.0.1",
     "nativescript-social-share": "1.3.1",
     "nativescript-telerik-ui-pro": "1.4.1",
@@ -51,20 +54,22 @@
   },
   "devDependencies": {
     "@types/chai": "3.4.34",
-    "babel-traverse": "6.8.0",
-    "babel-types": "6.8.1",
-    "babylon": "6.8.0",
+    "babel-traverse": "6.16.0",
+    "babel-types": "6.16.0",
+    "babylon": "6.11.4",
     "codelyzer": "0.0.28",
-    "filewalker": "0.1.2",
-    "jasmine-core": "2.4.1",
-    "karma": "1.2.0",
+    "filewalker": "0.1.3",
+    "jasmine-core": "2.5.2",
+    "karma": "1.3.0",
     "karma-jasmine": "1.0.2",
     "karma-nativescript-launcher": "0.4.0",
     "lazy": "1.0.11",
     "nativescript-dev-typescript": "0.3.2",
     "nativescript-dev-webpack": "0.0.20",
+    "rimraf": "2.5.4",
     "tslint": "3.15.1",
     "typed-graphql": "1.0.2",
-    "typescript": "2.0.3"
+    "typescript": "2.0.3",
+    "webpack": "1.13.2"
   }
 }
