@@ -2,6 +2,6 @@ import { Component } from "@angular/core";
 
 @Component({
   selector: "rom-main",
-  template: "<page-router-outlet></page-router-outlet>"
+  template: "<page-router-outlet></page-router-outlet>",
 })
 export class AppComponent { }
