@@ -1,15 +1,9 @@
-import { 
-  platformNativeScriptDynamic, 
-  NativeScriptModule 
-} from "nativescript-angular/platform";
-
 import { NgModule } from "@angular/core";
 import { NativeScriptRouterModule } from "nativescript-angular/router";
-import { NativeScriptFormsModule } from "nativescript-angular/forms";
-
+import { NativeScriptModule } from "nativescript-angular/platform";
 import ApolloClient, { createNetworkInterface } from "apollo-client";
 import { ApolloModule } from "angular2-apollo";
-
+import { NativeScriptServicesModule } from "./nativescript-services";
 import { authProviders, appRoutes } from "./app.routes";
 import { AppComponent } from "./app.component";
 
@@ -18,16 +12,21 @@ import {
   LoginService,
   OverviewService,
   BluetoothService,
+  UserService,
 } from "./shared";
 
 import { LoginModule } from "./login/Login.module";
+import { NewUserModule } from "./new-user/new-user.module";
 import { OverviewModule } from "./overview/overview.module";
 import { BluetoothModule } from "./bluetooth/bluetooth.module";
 import { BluetoothDetailsModule } from "./bluetooth-details/bluetooth-details.module";
 
+declare var process: any;
+const IP = process.env.IP;
+
 const client = new ApolloClient({
-  networkInterface: createNetworkInterface("http://192.168.1.46:8080/graphql"),
-})
+  networkInterface: createNetworkInterface(`http://${IP}:8080/graphql`),
+});
 
 setStatusBarColors();
 
@@ -37,18 +36,21 @@ setStatusBarColors();
     OverviewService,
     BluetoothService,
     authProviders,
+    UserService,
   ],
   imports: [
+    LoginModule,
     NativeScriptModule,
     NativeScriptRouterModule,
     NativeScriptRouterModule.forRoot(appRoutes),
-    LoginModule,
+    NativeScriptServicesModule,
+    NewUserModule,
     OverviewModule,
     BluetoothModule,
     BluetoothDetailsModule,
-    ApolloModule.withClient(client)
+    ApolloModule.withClient(client),
   ],
   declarations: [AppComponent],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule { }
