@@ -1,9 +1,7 @@
-import { Component, ElementRef, OnInit, ViewChild } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
-import { Animation } from "ui/animation";
 import { Page } from "ui/page";
-import { connectionType, getConnectionType } from "connectivity";
-import { prompt } from "ui/dialogs";
+import { DialogService } from "../nativescript-services";
 import { ItemEventData } from "ui/list-view";
 import { ObservableArray } from "data/observable-array";
 import { BluetoothService } from "../shared/bluetooth.service";
@@ -16,33 +14,46 @@ import * as bluetooth from "nativescript-bluetooth";
   styleUrls: ["bluetooth/bluetooth-common.css"],
 })
 export class BluetoothComponent implements OnInit {
-  peripherals: ObservableArray<bluetooth.Peripheral>;
-  isScanning = false;
+  public peripherals: ObservableArray<bluetooth.Peripheral>;
+  public isScanning = false;
 
   constructor(
     private router: Router,
     private page: Page,
-    private bluetoothService: BluetoothService
+    private bluetoothService: BluetoothService,
+    private dialogService: DialogService
   ) {
     this.peripherals = new ObservableArray<bluetooth.Peripheral>();
   }
 
-  ngOnInit() {
+  public ngOnInit() {
     this.page.actionBarHidden = true;
   }
 
+  /**
+   * Scan after Peripherals.
+   */
   public scan(): void {
     this.isScanning = true;
     this.peripherals.length = 0;
-    this.bluetoothService.scan( (peripheral: bluetooth.Peripheral) =>  {
-      this.peripherals.push(peripheral);
-    }).catch( () => {
-      // TODO: some error handling
+    this.bluetoothService.isBluetoothEnabled().then( (enabled) => {
+      if (!enabled) {
+        this.dialogService.alert("bluetooth is not enabled", "Bluetooth");
+        return;
+      }
+      this.bluetoothService.scan((peripheral: bluetooth.Peripheral) => {
+        this.peripherals.push(peripheral);
+      }).catch((err) => {
+        this.dialogService.alert(err, "Bluetooth");
+      });
     }).then( () => {
       this.isScanning = false;
     });
   }
 
+  /**
+   * Handle Peripheral tap.
+   */
   public onPeripheralTap(event: ItemEventData): void {
     const peripheral = this.peripherals.getItem(event.index);
     // TODO: SAVE UUID ON SESSION SERVICE
