@@ -1,16 +1,14 @@
 import { Component, ElementRef, OnInit, ViewChild } from "@angular/core";
 import { Router } from "@angular/router";
 import { Color } from "color";
-import { connectionType, getConnectionType } from "connectivity";
 import { Animation } from "ui/animation";
 import { View } from "ui/core/view";
-import { prompt } from "ui/dialogs";
 import { Page } from "ui/page";
 import { TextField } from "ui/text-field";
-import { Angular2Apollo } from "angular2-apollo";
-import gql from "graphql-tag";
+import { GridLayout } from "ui/layouts/grid-layout";
 
-import { alert, setHintColor, LoginService, User } from "../shared";
+import { ConnectivityService, DialogService } from "../nativescript-services";
+import { setHintColor, LoginService, Login } from "../shared";
 
 @Component({
   selector: "rom-login",
@@ -18,54 +16,106 @@ import { alert, setHintColor, LoginService, User } from "../shared";
   styleUrls: ["login/login-common.css", "login/login.component.css"],
 })
 export class LoginComponent implements OnInit {
-  user: User;
-  isLoggingIn = true;
-  isAuthenticating = false;
+  public user: Login;
+  public isLoggingIn = true;
+  public isAuthenticating = false;
 
-  @ViewChild("initialContainer") initialContainer: ElementRef;
-  @ViewChild("mainContainer") mainContainer: ElementRef;
-  @ViewChild("logoContainer") logoContainer: ElementRef;
-  @ViewChild("formControls") formControls: ElementRef;
-  @ViewChild("signUpStack") signUpStack: ElementRef;
-  @ViewChild("email") email: ElementRef;
-  @ViewChild("password") password: ElementRef;
+  @ViewChild("mainContainer") public mainContainer: ElementRef;
+  @ViewChild("formControls") public formControls: ElementRef;
+  @ViewChild("signUpStack") public signUpStack: ElementRef;
+  @ViewChild("email") public email: ElementRef;
+  @ViewChild("password") public password: ElementRef;
+  @ViewChild("background") public background: ElementRef;
 
   constructor(
-    private angularApollo: Angular2Apollo,
     private router: Router,
+    private page: Page,
     private loginService: LoginService,
-    private page: Page
+    private connectivityService: ConnectivityService,
+    private dialogService: DialogService
   ) {
-    this.user = new User();
+    this.user = new Login();
     this.user.email = "johnny@test.dk";
     this.user.password = "1234";
   }
 
-  ngOnInit() {
+  public ngOnInit() {
     this.page.actionBarHidden = true;
+    this.showMainContent();
   }
 
-  focusPassword() {
+  public newUser(): void {
+    this.router.navigate(["newUser"]);
+  }
+
+  public focusPassword() {
     this.password.nativeElement.focus();
   }
 
-  submit() {
+  public submit() {
     if (!this.user.isValidEmail()) {
-      alert("Enter a valid email address.");
+      this.dialogService.alert("Enter a valid email address.");
       return;
     }
 
-    this.isAuthenticating = true;
     if (this.isLoggingIn) {
+      this.isAuthenticating = true;
       this.login();
     } else {
       this.signUp();
     }
   }
 
-  login() {
-    if (getConnectionType() === connectionType.none) {
-      alert("Racket'O'Meter requires an internet connection to log in.");
+  public forgotPassword() {
+    if (this.isAuthenticating) {
+      return;
+    }
+
+    this.dialogService.prompt({
+      title: "Forgot Password",
+      message: "Enter the email address you used to register for Groceries to reset your password.",
+      okButtonText: "Ok",
+      cancelButtonText: "Cancel",
+    }).then((data) => {
+      if (data.result) {
+        this.loginService.resetPassword(data.text.trim())
+          .then(() => {
+            this.dialogService
+              .alert("Your password was successfully reset. Please check your email for instructions on choosing a new password.");
+          })
+          .catch(() => {
+            this.dialogService.alert("Unfortunately, an error occurred resetting your password.");
+          });
+      }
+    });
+  }
+
+  public toggleDisplay() {
+    if (this.isAuthenticating) {
+      return;
+    }
+
+    this.isLoggingIn = !this.isLoggingIn;
+    this.setTextFieldColors();
+    const mainContainer = this.mainContainer.nativeElement as View;
+    mainContainer.animate({
+      backgroundColor: this.isLoggingIn ? new Color("white") : new Color("#220c08"),
+      duration: 200,
+    });
+  }
+
+  public startBackgroundAnimation() {
+    const background = this.background.nativeElement as GridLayout;
+
+    background.animate({
+      scale: { x: 1.0, y: 1.0 },
+      duration: 10000,
+    });
+  }
+
+  private login() {
+    if (!this.connectivityService.isOnline()) {
+      this.dialogService.alert("Racket'O'Meter requires an internet connection to log in.");
       return;
     }
 
@@ -77,112 +127,50 @@ export class LoginComponent implements OnInit {
         }
       })
       .catch((message) => {
-        alert("message: " + message);
+        this.dialogService.alert("message: " + message);
         this.isAuthenticating = false;
       });
   }
 
-  signUp() {
-    if (getConnectionType() === connectionType.none) {
-      alert("Groceries requires an internet connection to register.");
+  private signUp() {
+    if (!this.connectivityService.isOnline()) {
+      this.dialogService.alert("Internet is requried to register.");
       return;
     }
 
-    this.loginService.register(this.user)
-      .then(() => {
-        alert("Your account was successfully created.");
-        this.isAuthenticating = false;
-        this.toggleDisplay();
-      })
-      .catch((message) => {
-        if (message.match(/same user/)) {
-          alert("This email address is already in use.");
-        } else {
-          alert("Unfortunately we were unable to create your account.");
-        }
-        this.isAuthenticating = false;
-      });
+    this.dialogService.alert("Registration is not possible at the time.");
   }
 
-  forgotPassword() {
-    prompt({
-      title: "Forgot Password",
-      message: "Enter the email address you used to register for Groceries to reset your password.",
-      defaultText: "",
-      okButtonText: "Ok",
-      cancelButtonText: "Cancel"
-    }).then((data) => {
-      if (data.result) {
-        this.loginService.resetPassword(data.text.trim())
-          .then(() => {
-            alert("Your password was successfully reset. Please check your email for instructions on choosing a new password.");
-          })
-          .catch(() => {
-            alert("Unfortunately, an error occurred resetting your password.");
-          });
-      }
-    });
+  private showMainContent() {
+    const mainContainer = this.mainContainer.nativeElement as View;
+    const formControls = this.formControls.nativeElement as View;
+    const signUpStack = this.signUpStack.nativeElement as View;
+    const animations = [];
+
+    // Show the main container. The main container will
+    // not immediately appear because their opacity is set to 0 in CSS.
+    mainContainer.style.visibility = "visible";
+
+    // Fade in the main container over one half second.
+    animations.push({ target: mainContainer, opacity: 1, duration: 500 });
+
+    // Slide up the form controls and sign up container.
+    animations.push({ target: signUpStack, translate: { x: 0, y: 0 }, opacity: 1, delay: 500, duration: 150 });
+    animations.push({ target: formControls, translate: { x: 0, y: 0 }, opacity: 1, delay: 650, duration: 150 });
+
+    // Kick off the animation queue
+    new Animation(animations, false).play();
   }
 
-  toggleDisplay() {
-    this.isLoggingIn = !this.isLoggingIn;
-    this.setTextFieldColors();
-    let mainContainer = <View>this.mainContainer.nativeElement;
-    mainContainer.animate({
-      backgroundColor: this.isLoggingIn ? new Color("white") : new Color("#301217"),
-      duration: 200
-    });
-  }
+  private setTextFieldColors() {
+    const emailTextField = this.email.nativeElement as TextField;
+    const passwordTextField = this.password.nativeElement as TextField;
 
-  startBackgroundAnimation(background) {
-    background.animate({
-      scale: { x: 1.0, y: 1.0 },
-      duration: 10000
-    });
-  }
-
-  showMainContent() {
-    let initialContainer = <View>this.initialContainer.nativeElement;
-    let mainContainer = <View>this.mainContainer.nativeElement;
-    let logoContainer = <View>this.logoContainer.nativeElement;
-    let formControls = <View>this.formControls.nativeElement;
-    let signUpStack = <View>this.signUpStack.nativeElement;
-    let animations = [];
-
-    // Fade out the initial content over one half second
-    initialContainer.animate({
-      opacity: 0,
-      duration: 500
-    }).then(function () {
-      // After the animation completes, hide the initial container and
-      // show the main container and logo. The main container and logo will
-      // not immediately appear because their opacity is set to 0 in CSS.
-      initialContainer.style.visibility = "collapse";
-      mainContainer.style.visibility = "visible";
-      logoContainer.style.visibility = "visible";
-
-      // Fade in the main container and logo over one half second.
-      animations.push({ target: mainContainer, opacity: 1, duration: 500 });
-      animations.push({ target: logoContainer, opacity: 1, duration: 500 });
-
-      // Slide up the form controls and sign up container.
-      animations.push({ target: signUpStack, translate: { x: 0, y: 0 }, opacity: 1, delay: 500, duration: 150 });
-      animations.push({ target: formControls, translate: { x: 0, y: 0 }, opacity: 1, delay: 650, duration: 150 });
-
-      // Kick off the animation queue
-      new Animation(animations, false).play();
-    });
-  }
-
-  setTextFieldColors() {
-    let emailTextField = <TextField>this.email.nativeElement;
-    let passwordTextField = <TextField>this.password.nativeElement;
-
-    let mainTextColor = new Color(this.isLoggingIn ? "black" : "#C4AFB4");
+    const mainTextColor = new Color(this.isLoggingIn ? "black" : "#C4AFB4");
     emailTextField.color = mainTextColor;
     passwordTextField.color = mainTextColor;
 
-    let hintColor = new Color(this.isLoggingIn ? "#ACA6A7" : "#C4AFB4");
+    const hintColor = new Color(this.isLoggingIn ? "#ACA6A7" : "#C4AFB4");
     setHintColor({ view: emailTextField, color: hintColor });
     setHintColor({ view: passwordTextField, color: hintColor });
   }
