@@ -1,4 +1,4 @@
-import { platformNativeScriptDynamic, NativeScriptModule } from "nativescript-angular/platform";
+import { platformNativeScriptDynamic } from "nativescript-angular/platform";
 import { AppModule } from "./app.module";
 
 platformNativeScriptDynamic().bootstrapModule(AppModule);
