@@ -2,6 +2,7 @@ export * from "./user.model";
 export * from "./login.service";
 export * from "./bluetooth.service";
 export * from "./overview.service";
-export * from "./dialog-util";
 export * from "./hint-util";
 export * from "./status-bar-util";
+export * from "./login.model";
+export * from "./user.service";
